refactor: Refactor the service of retrieval the recommend app
