+++ api/services/recommend_app/recommend_app_type.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class RecommendAppType(str, Enum):
+    REMOTE = "remote"
+    BUILDIN = "builtin"
+    DATABASE = "db"
